added cloudinary cloud for images

+++ controllers/foodControllers.js
@@ -1,24 +1,28 @@
 import foodModel from "../models/foodModels.js";
-import fs from "fs";
+import cloudinary from "../Config/cloudinaryConfig.js";
 
 // Add food Item
 
 const addFood = async (req, res) => {
-  let image_filename = `${req.file.filename}`;
-
-  const food = new foodModel({
-    name: req.body.name,
-    description: req.body.description,
-    price: req.body.price,
-    category: req.body.category,
-    image: image_filename,
-  });
   try {
+    const result = await cloudinary.uploader.upload(req.file.path, {
+      folder: "foodItems",
+      use_filename: true,
+    });
+
+    const food = new foodModel({
+      name: req.body.name,
+      description: req.body.description,
+      price: req.body.price,
+      category: req.body.category,
+      image: result.secure_url, // Save the Cloudinary URL
+    });
+
     await food.save();
-    res.json({ success: true, message: "Food Added" });
+    res.json({ success: true, message: "Food Added", data: food });
   } catch (error) {
     console.log(error);
-    res.json({ success: false, message: "Error" });
+    res.json({ success: false, message: "Error Adding Food" });
   }
 };
 
@@ -34,17 +38,22 @@ const listFood = async (req, res) => {
 };
 
 //Remove Food Item
-
 const removeFood = async (req, res) => {
   try {
     const food = await foodModel.findById(req.body.id);
+    if (!food) {
+      return res.json({ success: false, message: "Food Item Not Found" });
+    }
 
-    fs.unlink(`Uploads/${food.image}`, () => {}); // Delete Images from uploads and mongoDB
+    const publicId = food.image.split("/").pop().split(".")[0];
+
+    await cloudinary.uploader.destroy(`foodItems/${publicId}`);
     await foodModel.findByIdAndDelete(req.body.id);
-    res.json({ success: true, mesaage: "Food Successfully Removed" });
+
+    res.json({ success: true, message: "Food Successfully Removed" });
   } catch (error) {
     console.log(error);
-    res.json({ success: false, message: "Error have been founded " });
+    res.json({ success: false, message: "Error Removing Food Item" });
   }
 };
